Add new testframe work, in progress
