--- unit_tests/run_all
@@ -1,32 +1,0 @@
-#!/usr/bin/env python3
-"""
-Unit tests.
-"""
-
-from unit_test_common import initialize_csv2_request
-import sys
-
-import user_commands
-import user_requests
-import user_requests_setup
-import user_requests_cleanup
-
-def main():
-    gvar = {}
-    if len(sys.argv) > 1:
-        initialize_csv2_request(gvar, sys.argv[0], selections=sys.argv[1])
-    else:
-        initialize_csv2_request(gvar, sys.argv[0])
-
-    user_commands.main(gvar)
-
-    user_requests_setup.main(gvar)
-    user_requests.main(gvar)
-    user_requests_cleanup.main(gvar)
-
-    print('Test run=%s, skipped=%s, failed=%s.' % (gvar['ut_count']-gvar['ut_skipped'], gvar['ut_skipped'], gvar['ut_failed']))
-    exit(gvar['ut_failed'])
-
-if __name__ == "__main__":
-    main()
-
